fixed: tessdata not bundled with setup option 1 only

+++ Setup/src/main/java/org/sikuli/setup/RunSetup.java
@@ -909,7 +909,7 @@ public class RunSetup {
       success &= handleTempAfter(targetJar, localJar);
     }
 
-    if(getTess) {
+    if(getAPI && getTess) {
       new File(workDir, localTess).delete();
       jarsList[2] = null;
     }
